Adds core_file traversals to traversals module
Fixes #8.  Previously, these two methods (x#content_objects and
x#canonical_object) weren't implemented on the Traversals module for
fear that they muddied the water by existing on Collections and
Communities.  Given that literally every collection-only/community-only
traversal method is already implemented on the CoreFile class this was a
bit of a silly restriction.  Future changes may ensure that these
methods are provided in a more restricted fashion.