Change default query behavior and arg passing
Instead of defaulting to returning the raw response hash from
ActiveFedora's SolrQueryService, all methods in the Traversals concern
now default to returning ORM models.  This changes the default behavior
from doing what is fastest to doing what returns the most immediately
useful object for most cases.

Furthermore, return type is no longer specified as an option in a hash.
Instead it is passed as a single argument that has a default value of
:models in the Traversals concern.  This is considerably less verbose
and not in any real sense (at least not currently) less flexible.

Fixes specs to reflect these changes.